EP-3612 move S1BackscatterOrfeo to own module for cleaner overview
